+++ src/_structures_mod.jl
@@ -5,8 +5,7 @@ function fit!(mod::Jchemo.Transformer, X)
     mod.fm = mod.fun(X; kwargs...)
     return
 end  
-function fit!(mod::Jchemo.Transformer, X, 
-        weights::Weight)
+function fit!(mod::Jchemo.Transformer, X, weights::Weight)
     kwargs = values(mod.kwargs)
     mod.fm = mod.fun(X, weights; kwargs...)
     return
@@ -17,8 +16,7 @@ function fit!(mod::Jchemo.Transformer, X, Y)
     mod.fm = mod.fun(X, Y; kwargs...)
     return
 end  
-function fit!(mod::Jchemo.Transformer, X, Y, 
-        weights::Weight)
+function fit!(mod::Jchemo.Transformer, X, Y, weights::Weight)
     kwargs = values(mod.kwargs)
     mod.fm = mod.fun(X, Y, weights; kwargs...)
     return
@@ -29,8 +27,7 @@ function fit!(mod::Jchemo.Predictor, X, Y)
     mod.fm = mod.fun(X, Y; kwargs...)
     return
 end  
-function fit!(mod::Jchemo.Predictor, X, Y, 
-        weights::Weight)
+function fit!(mod::Jchemo.Predictor, X, Y, weights::Weight)
     kwargs = values(mod.kwargs)
     mod.fm = mod.fun(X, Y, weights; kwargs...)
     return
@@ -42,34 +39,29 @@ function fit!(mod::PredictorNoY, X)
 end   
 
 ######---- Transf
-function transf(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, 
-        X; nlv = nothing)
+function transf(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, X; nlv = nothing)
     isnothing(nlv) ? transf(mod.fm, X) : 
         transf(mod.fm, X; nlv)
 end
 ## X, Y
-function transf(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, 
-        X, Y; nlv = nothing)
+function transf(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, X, Y; nlv = nothing)
     isnothing(nlv) ? transf(mod.fm, X, Y) : 
         transf(mod.fm, X, Y; nlv)
 end
 
 ######---- Transfbl
-function transfbl(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, 
-        X; nlv = nothing)
+function transfbl(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, X; nlv = nothing)
     isnothing(nlv) ? transfbl(mod.fm, X) : 
         transfbl(mod.fm, X; nlv)
 end
 ## X, Y
-function transfbl(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, 
-        X, Y; nlv = nothing)
+function transfbl(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, X, Y; nlv = nothing)
     isnothing(nlv) ? transfbl(mod.fm, X, Y) : 
     transfbl(mod.fm, X, Y; nlv)
 end
 
 ######---- Predict 
-function predict(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, 
-        X; nlv = nothing, lb = nothing)
+function predict(mod::Union{Jchemo.Transformer, Jchemo.Predictor}, X; nlv = nothing, lb = nothing)
     if isnothing(nlv) && isnothing(lb)
         predict(mod.fm, X)
     elseif !isnothing(nlv) 
@@ -81,8 +73,7 @@ end
 predict(mod::PredictorNoY, X) = predict(mod.fm, X)
 
 ######---- Coef 
-function coef(mod::Jchemo.Predictor; 
-        nlv = nothing, lb = nothing)
+function coef(mod::Jchemo.Predictor; nlv = nothing, lb = nothing)
     if isnothing(nlv) && isnothing(lb)
         coef(mod.fm)
     elseif !isnothing(nlv)