+++ src/_structures_mod_pipelines.jl
@@ -7,7 +7,7 @@ pip(args...) = Pipeline(values(args))
 function fit!(mod::Pipeline, X, Y = nothing)
     K = length(mod.mod)
     @assert K > 1 "Wrong pipeline: must contain at least 2 models."
-    for i = 1:(K - 1)
+    @inbounds for i = 1:(K - 1)
         if isa(mod.mod[i], Transformer)
             fit!(mod.mod[i], X)
         elseif isa(mod.mod[i], Predictor)
@@ -26,7 +26,7 @@ end
 function transf(mod::Pipeline, X)  
     K = length(mod.mod)
     @assert K > 1 "Wrong pipeline: must contain at least 2 models."
-    for i = 1:K
+    @inbounds for i = 1:K
         X = transf(mod.mod[i], X)
     end
     X
@@ -36,7 +36,7 @@ end
 function predict(mod::Pipeline, X)
     K = length(mod.mod)
     @assert K > 1 "Wrong pipeline: must contain at least 2 models."
-    for i = 1:(K - 1)
+    @inbounds for i = 1:(K - 1)
         X = transf(mod.mod[i], X)
     end
     predict(mod.mod[K], X)